requirements.txt
Django==3.1.1
newspaper3k==0.2.8
pre-commit==2.7.1
python-dotenv==0.14.0
django-npm==1.0.0
